requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
spotipy
requests
matplotlib
Pillow
